Expt-003-fourth-power

+++ data.h
@@ -1,18 +1,18 @@
 extern int shouldquit;
 
-typedef struct { float probReproK0; float probReproK1; float probReproK2; } Code;
+typedef struct { double probReproK0; double probReproK1; double probReproK2; } Code;
 
-typedef struct { float bal; Code code; } Animal;
+typedef struct { double bal; Code code; } Animal;
 
 DECLARE_LISTED(Animal)
 
 typedef struct { 
 	ListedAnimal * animals; 
 	int numAnimals;
-	float avgProbReproK0;
-	float avgProbReproK1;
-	float avgProbReproK2;
-	float avgBal;
+	double avgProbReproK0;
+	double avgProbReproK1;
+	double avgProbReproK2;
+	double avgBal;
 } World;
 
 extern int iteration;
